+++ packages_flutter/sqflite_ffi/lib/src/sqflite_ffi.dart
@@ -1,0 +1,2 @@
+export 'sqflite_ffi_io.dart'
+    if (dart.library.js_interop) 'sqflite_ffi_web.dart';